+++ src/app/(dashboard)/dashboard/p/[id]/layout.tsx
