+++ src/app/(dashboard)/dashboard/p/[id]/settings/page.tsx
@@ -1,0 +1,13 @@
+import React, { FC } from 'react'
+
+interface pageProps {}
+
+const page: FC<pageProps> = async ({}) => {
+   return (
+       <>
+           <div>Project Settings</div>
+       </>
+   )
+}
+
+export default page
